requirements.txt
Flask==3.1.0
requests==2.32.3
python-dotenv==1.2.2
newspaper3k==0.2.8
lxml[html_clean]